+++ src/components/account/forms/FindPasswordForm.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { useForm } from "react-hook-form";
 import { Input, Label } from "../../../styles/AccountStyles";
 import { IFindPasswordForm } from "../../../interfaces/form";
-import { findPasswordPost } from "../../../api/accountApi";
+import { findPasswordPost } from "../../../apis/api/accountApi";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
