+++ src/components/account/forms/SignUpForm.tsx
@@ -4,7 +4,7 @@ import { Input, Label } from "../../../styles/AccountStyles";
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { ISignupForm } from "../../../interfaces/form";
-import { idCheckPost, signupPost } from "../../../api/accountApi";
+import { idCheckPost, signupPost } from "../../../apis/api/accountApi";
 import { getValue } from "@testing-library/user-event/dist/utils";
 import { sign } from "crypto";
 import React from "react";
